+++ packages/Wallet/Core/database/migrations/2021_11_18_043024_create_transaction_charges_table.php
@@ -16,9 +16,9 @@ class CreateTransactionChargesTable extends Migration
         Schema::create('transaction_charges', function (Blueprint $table) {
             $table->id();
             $table->foreignId('payment_channel_id')->nullable()->references("id")->on("payment_channels")->onDelete("set null");
-            $table->double("minimum")->default(0);
-            $table->double("maximum")->default(0);
-            $table->double("charge")->default(0);
+            $table->decimal("minimum", 10, 2)->default(0);
+            $table->decimal("maximum", 10, 2)->default(0);
+            $table->decimal("charge", 10, 2)->default(0);
             $table->timestamps();
         });
     }
